requirements.txt
dj-database-url==0.5.0
Django==2.2.7
django-heroku==0.3.1
django-service-objects==0.5.0
djangorestframework==3.11.0
pkg-resources==0.0.0
psycopg2==2.8.4
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
whitenoise==5.0.1
